requirements.txt
<<<<<<< HEAD
=======
# FastAPI backend
fastapi
uvicorn
python-dotenv

>>>>>>> backend
# Google / Gmail API
google-auth
google-auth-oauthlib
google-api-python-client

# PDF reading
pdfplumber
pdfminer.six

<<<<<<< HEAD
# word reading
python-docx

#excel reading
openpyxl

#image reading
pytesseract pillow

# Text-to-speech
pyttsx3
pypiwin32; platform_system == "Windows"

# Ollama HTTP API
=======
# Word reading
python-docx

# Excel reading
openpyxl

# Image reading (works locally, limited on Render)
pytesseract
pillow

# HTTP client (for LLM API replacement later)
>>>>>>> backend
requests
